requirements.txt
mcp[cli]>=1.0.0
ollama>=0.1.10
bcrypt>=4.0.0
httpx>=0.27.0
uvicorn>=0.24.0
starlette>=0.27.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
